# docs/app/swagger.yaml
basePath: /api/v1
definitions:
...
  placio-app_Dto.Error:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  placio-app_Dto.ErrorDTO:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  placio-app_Dto.EventDTO:
    properties:
      coverImage:
        type: string
      createdAt:
        type: string
      description:
        type: string
      endDate:
        type: string
      endTime:
        type: string
      eventSettings:
        additionalProperties: true
        type: object
      eventType:
        type: string
      frequency:
        type: string
      frequencyDayOfMonth:
        type: string
      frequencyDayOfWeek:
        type: string
      frequencyInterval:
        type: string
      frequencyMonthOfYear:
        type: string
      id:
        type: string
      location:
        type: string
      name:
        type: string
      startDate:
        type: string
      startTime:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      timeZone:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      venueAddress:
        type: string
      venueCity:
        type: string
      venueCountry:
        type: string
      venueEmail:
        type: string
      venueLat:
        type: string
      venueLon:
        type: string
      venueName:
        type: string
      venuePhone:
        type: string
      venueState:
        type: string
      venueType:
        type: string
      venueURL:
        type: string
      venueZIP:
        type: string
    type: object
...
  placio-app_ent.Business:
    properties:
      business_settings:
        additionalProperties: true
        description: BusinessSettings holds the value of the "business_settings" field.
        type: object
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.BusinessEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BusinessQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      picture:
        description: Picture holds the value of the "picture" field.
        type: string
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      url:
        description: URL holds the value of the "url" field.
        type: string
      website:
        description: Website holds the value of the "website" field.
        type: string
    type: object
...
  placio-app_ent.BusinessFollowEvent:
    properties:
      createdAt:
        description: CreatedAt holds the value of the "createdAt" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.BusinessFollowEventEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BusinessFollowEventQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      updatedAt:
        description: UpdatedAt holds the value of the "updatedAt" field.
        type: string
    type: object
...
  placio-app_ent.Category:
    properties:
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.CategoryEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CategoryQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      image:
        description: Image holds the value of the "image" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  placio-app_ent.CategoryAssignment:
    properties:
      category_id:
        description: This represents the ID of the category
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.CategoryAssignmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CategoryAssignmentQuery when eager-loading is set.
      entity_id:
        description: This represents the ID of User, Business or Place entity
        type: string
      entity_type:
        description: 'This represents the type of entity: User, Business, or Place'
        type: string
      id:
        description: ID of the ent.
        type: string
    type: object
...
  placio-app_ent.Event:
    properties:
      EventType:
        allOf:
        - $ref: '#/definitions/placio-app_ent_event.EventType'
        description: EventType holds the value of the "EventType" field.
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      createdAt:
        description: CreatedAt holds the value of the "createdAt" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.EventEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EventQuery when eager-loading is set.
      end_date:
        description: EndDate holds the value of the "end_date" field.
        type: string
      end_time:
        description: EndTime holds the value of the "end_time" field.
        type: string
      event_settings:
        additionalProperties: true
        description: EventSettings holds the value of the "event_settings" field.
        type: object
      frequency:
        allOf:
        - $ref: '#/definitions/placio-app_ent_event.Frequency'
        description: Frequency holds the value of the "frequency" field.
      frequency_day_of_month:
        description: FrequencyDayOfMonth holds the value of the "frequency_day_of_month"
          field.
        type: string
      frequency_day_of_week:
        description: FrequencyDayOfWeek holds the value of the "frequency_day_of_week"
          field.
        type: string
      frequency_interval:
        description: FrequencyInterval holds the value of the "frequency_interval"
          field.
        type: string
      frequency_month_of_year:
        description: FrequencyMonthOfYear holds the value of the "frequency_month_of_year"
          field.
        type: string
      id:
        description: ID of the ent.
        type: string
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      start_date:
        description: StartDate holds the value of the "start_date" field.
        type: string
      start_time:
        description: StartTime holds the value of the "start_time" field.
        type: string
      status:
        description: Status holds the value of the "status" field.
        type: string
      tags:
        description: Tags holds the value of the "tags" field.
        type: string
      time_zone:
        description: TimeZone holds the value of the "time_zone" field.
        type: string
      title:
        description: Title holds the value of the "title" field.
        type: string
      updatedAt:
        description: UpdatedAt holds the value of the "updatedAt" field.
        type: string
      url:
        description: URL holds the value of the "url" field.
        type: string
      venue_address:
        description: VenueAddress holds the value of the "venue_address" field.
        type: string
      venue_city:
        description: VenueCity holds the value of the "venue_city" field.
        type: string
      venue_country:
        description: VenueCountry holds the value of the "venue_country" field.
        type: string
      venue_email:
        description: VenueEmail holds the value of the "venue_email" field.
        type: string
      venue_lat:
        description: VenueLat holds the value of the "venue_lat" field.
        type: string
      venue_lon:
        description: VenueLon holds the value of the "venue_lon" field.
        type: string
      venue_name:
        description: VenueName holds the value of the "venue_name" field.
        type: string
      venue_phone:
        description: VenuePhone holds the value of the "venue_phone" field.
        type: string
      venue_state:
        description: VenueState holds the value of the "venue_state" field.
        type: string
      venue_type:
        allOf:
        - $ref: '#/definitions/placio-app_ent_event.VenueType'
        description: VenueType holds the value of the "venue_type" field.
      venue_url:
        description: VenueURL holds the value of the "venue_url" field.
        type: string
      venue_zip:
        description: VenueZip holds the value of the "venue_zip" field.
        type: string
    type: object
  placio-app_ent.EventEdges:
    properties:
      businessFollowers:
        description: BusinessFollowers holds the value of the businessFollowers edge.
        items:
          $ref: '#/definitions/placio-app_ent.BusinessFollowEvent'
        type: array
      event_categories:
        description: EventCategories holds the value of the event_categories edge.
        items:
          $ref: '#/definitions/placio-app_ent.Category'
        type: array
      event_category_assignments:
        description: EventCategoryAssignments holds the value of the event_category_assignments
          edge.
        items:
          $ref: '#/definitions/placio-app_ent.CategoryAssignment'
        type: array
      faqs:
        description: Faqs holds the value of the faqs edge.
        items:
          $ref: '#/definitions/placio-app_ent.FAQ'
        type: array
      ownerBusiness:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Business'
        description: OwnerBusiness holds the value of the ownerBusiness edge.
      ownerUser:
        allOf:
        - $ref: '#/definitions/placio-app_ent.User'
        description: OwnerUser holds the value of the ownerUser edge.
      place:
        description: Place holds the value of the place edge.
        items:
          $ref: '#/definitions/placio-app_ent.Place'
        type: array
      ratings:
        description: Ratings holds the value of the ratings edge.
        items:
          $ref: '#/definitions/placio-app_ent.Rating'
        type: array
      ticket_options:
        description: TicketOptions holds the value of the ticket_options edge.
        items:
          $ref: '#/definitions/placio-app_ent.TicketOption'
        type: array
      tickets:
        description: Tickets holds the value of the tickets edge.
        items:
          $ref: '#/definitions/placio-app_ent.Ticket'
        type: array
      userFollowers:
        description: UserFollowers holds the value of the userFollowers edge.
        items:
          $ref: '#/definitions/placio-app_ent.UserFollowEvent'
        type: array
    type: object
  placio-app_ent.FAQ:
    properties:
      answer:
        description: Answer holds the value of the "answer" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.FAQEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FAQQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      question:
        description: Question holds the value of the "question" field.
        type: string
    type: object
...
  placio-app_ent.Rating:
    properties:
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.RatingEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RatingQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      ratedAt:
        description: When the rating was given.
        type: string
      review:
        description: User's review to the business/place/event.
        type: string
      score:
        description: Score should be between 1 and 5.
        type: integer
    type: object
...
  placio-app_ent_event.EventType:
    enum:
    - event
    - place
    - business
    - free
    - paid
    type: string
    x-enum-varnames:
    - EventTypeEvent
    - EventTypePlace
    - EventTypeBusiness
    - EventTypeFree
    - EventTypePaid
  placio-app_ent_event.Frequency:
    enum:
    - once
    - daily
    - weekly
    - monthly
    - yearly
    type: string
    x-enum-varnames:
    - FrequencyOnce
    - FrequencyDaily
    - FrequencyWeekly
    - FrequencyMonthly
    - FrequencyYearly
  placio-app_ent_event.VenueType:
    enum:
    - online
    - in_person
    - hybrid
    type: string
    x-enum-varnames:
    - VenueTypeOnline
    - VenueTypeInPerson
    - VenueTypeHybrid
...
  placio-app_models.Attendee:
    properties:
      attended:
        type: boolean
      createdAt:
        type: string
      eventID:
        type: string
      id:
        type: string
      ticketID:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
...
  placio-app_models.Rating:
    properties:
      createdAt:
        type: string
      eventID:
        type: string
      id:
        type: string
      review:
        type: string
      score:
        type: integer
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  placio-app_models.Ticket:
    properties:
      attendees:
        items:
          $ref: '#/definitions/placio-app_models.Attendee'
        type: array
      createdAt:
        type: string
      eventID:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      ratings:
        description: Comments      []Comment      `gorm:"foreignKey:EventID"`
        items:
          $ref: '#/definitions/placio-app_models.Rating'
        type: array
      ticketOptions:
        items:
          $ref: '#/definitions/placio-app_models.TicketOption'
        type: array
      updatedAt:
        type: string
    type: object
  placio-app_models.TicketOption:
    properties:
      createdAt:
        type: string
      eventID:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: https://api.palnight.com
info:
  contact:
    email: support@placio.io
    name: Placio Ltd
    url: https://placio.io
  description: This is the documentation for the Placio Application Api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://placio.io/terms
  title: Placio Application Api
  version: "0.01"
paths:
...
  /api/v1/search/events:
    get:
      consumes:
      - application/json
      description: Search for Events by a given search text
      parameters:
      - description: Text to search for
        in: query
        name: searchText
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully found search results
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Search for Events
      tags:
      - Search
...
  /api/v1/tickets/event/{eventId}:
    get:
      consumes:
      - application/json
      description: GET all tickets for a specific event
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved tickets
          schema:
            items:
              $ref: '#/definitions/placio-app_models.Ticket'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: GET tickets by event ID
      tags:
      - Ticket
...
  /business/{businessAccountID}/event:
    post:
      consumes:
      - application/json
      description: Add a new Event to a Business Account
      operationId: add-a-new-event-to-business-account
      parameters:
      - description: Business Account ID
        in: path
        name: businessAccountID
        required: true
        type: string
      - description: Event DTO
        in: body
        name: Dto.EventDTO
        required: true
        schema:
          $ref: '#/definitions/placio-app_Dto.EventDTO'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_ent.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Add a new Event to a Business Account
      tags:
      - Business
...
  /events:
    get:
      consumes:
      - application/json
      description: Get Events
      parameters:
      - in: query
        items:
          type: string
        name: eventType
        type: array
      - in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        items:
          type: string
        name: location
        type: array
      - in: query
        items:
          type: string
        name: name
        type: array
      - in: query
        items:
          type: string
        name: status
        type: array
      - in: query
        items:
          type: string
        name: timeZone
        type: array
      - in: query
        items:
          type: string
        name: title
        type: array
      - in: query
        items:
          type: string
        name: url
        type: array
      - description: Page Number
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: pageSize
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Get all Events
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: Create Event
      parameters:
      - description: Event Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/placio-app_Dto.EventDTO'
      - description: Business ID
        in: query
        name: businessId
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_ent.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Create Event
      tags:
      - Event
  /events/{eventId}:
    delete:
      consumes:
      - application/json
      description: Delete an existing event
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Delete Event
      tags:
      - Event
    get:
      consumes:
      - application/json
      description: Get a single event by its ID
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_ent.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Get Event By ID
      tags:
      - Event
    put:
      consumes:
      - application/json
      description: Update Event
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Business ID
        in: query
        name: businessId
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/placio-app_Dto.EventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_ent.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Update Event
      tags:
      - Event
  /events/participants/{eventId}:
    get:
      consumes:
      - application/json
      description: Get Event Participants
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Get Event Participants
      tags:
      - Event
...
  /faq/{faqID}/event/{eventID}:
    post:
      consumes:
      - application/json
      description: Associate a FAQ with an Event by their IDs
      operationId: associate-FAQ-event
      parameters:
      - description: FAQ ID
        in: path
        name: faqID
        required: true
        type: string
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: FAQ associated with event successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Associate FAQ with Event
      tags:
      - FAQ
...
  /follow/event/{eventID}:
    post:
      consumes:
      - application/json
      description: Follow a specific Event by the user
      operationId: follow-event
      parameters:
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User successfully followed the event
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Follow an event
      tags:
      - Follow
...
  /followed/events:
    get:
      consumes:
      - application/json
      description: Get all Events followed by the user
      operationId: get-followed-events
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Get followed events by a user
      tags:
      - Follow
...
  /unfollow/event/{eventID}:
    delete:
      consumes:
      - application/json
      description: Unfollow a specific Event by the user
      operationId: unfollow-event
      parameters:
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User successfully unfollowed the event
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Unfollow an event
      tags:
      - Follow
...
schemes:
- http
- https
swagger: "2.0"